refactoring: agent class
